Class3/pyproject.toml
[project]
name = "class3"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "livekit-agents>=1.4.5",
    "livekit-plugins-cartesia>=1.4.5",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "livekit-plugins-openai>=1.4.5",
    "livekit-plugins-silero>=1.4.5",
    "livekit-plugins-turn-detector>=1.4.5",
    "python-dotenv>=1.2.2",
]
